Hide `=` from command handlers that use eq_sep_args ##newshell
Command handlers like `env`, `%` or `*` expect arguments with an equal
as a separator. So far `=` was visible as an arg_identifier to make sure
that existing handlers could still work. This is not required anymore,
as all commands that use eq_sep_args are now converted to a argv-style
handler.