Convert %, *, env, w6 commands to argv ##newshell (#16)
* No more FS plugins in rizin
* Fix retrieval of cmddesc with GROUP involved ##newshell
* Add w6d/w6e argv-style handlers ##newshell
* Add OLDINPUT_SPECIAL for special characters ##newshell
* Convert % to a argv command ##newshell
* Hide `=` from command handlers that use eq_sep_args ##newshell

Command handlers like `env`, `%` or `*` expect arguments with an equal
as a separator. So far `=` was visible as an arg_identifier to make sure
that existing handlers could still work. This is not required anymore,
as all commands that use eq_sep_args are now converted to a argv-style
handler.

* Convert * command to argv-style ##newshell