Use async-signal-safe functions after fork in multithread program (#207)
* Do not use setenv after fork, it's not async-signal-safe
* Do not use mutexes after fork

If pipe2 is not available on a system, we try to simulate the atomic
setting of O_CLOEXEC in different ways (depending on whether O_CLOEXEC
is at least available or not). However this involves using mutexes. When
forking, only the calling thread is forked, so it is important to lock
the mutex before the fork. However, in the child between the fork and
the exec* it is recommended to use only async-signal-safe functions. mutex
operations are not async-signal-safe, so we just keep track of whether we are
in the child or not and avoid using the mutex as necessary.

* Replace pthread_atfork with rz_sys_fork