Fix `???` command, which should show the help of `?` (#365)
* Fix `???` command, which should show the help of `?`

Due to recent changes it was not working anymore, because it was
interpreteing `???` as the help with detail=3 of the root node. This
patch limits the number of detail to 2 (which is what is already assumed
in another parts of the code anyway), so that `???` can be interpreted
as the help with detail=2 of `?`. For now this makes it work, however
there needs to be some review of the whole `?` subtree.

* Add test for ???

* Fix ??? test spaces

Co-authored-by: Anton Kochkov <anton.kochkov@gmail.com>
Co-authored-by: Florian Märkl <info@florianmaerkl.de>