core: automatically set LAST flag on last argument if of right type
Instead of relying on the developer to remember to use the
RZ_CMD_ARG_FLAG_LAST flag on the last argument when it is of the right
type, we can just set it automatically. We do it only for RZNUM, STRING
and CMD types and this can be bypassed if the devs really want. We do
not add this flag on all types because otherwise all commands handlers
would accept any number of argument. It is good instead to have the
parser immediately detect when too many arguments are passed to some
commands.