cons: avoid flushing stuff on the screen when push/pop are used
When you use rz_cons_push/pop you likely want to capture the whole
output between push/pop (e.g. to capture it as a string or because you
want to flush it all at the end).

This commit creates a new cons context with the `noflush` property set
to true, so that possible rz_cons_flush in command handlers do not
actually remove the text from the cons buffer out to the screen.

In particular, this fixes things like `(_;pd 1;?e)() @@f:flag` where the
macro handler is using rz_cons_flush and pd 1 output is directly printed
on the screen instead of being captured by iter_flags_command handler.