rizin-shell: do argument validation in the statement handlers
Move the validation of some arguments from the parser to the statement
handlers. This is necessary to allow commands substitutions to provide
more than one argument in some cases (e.g. `pd 10 @@s:$(?e 10 20 2)`).

If the check of the number of arguments is done in the parser, the above
command would be treated as having just one argument after `@@s:`.
However, when the command is substituted it actually provides 3
arguments. By moving the validation to the handlers, we can first
substitute the commands and then check if the number of arguments
provided is valid.