Fix undefined behaviour of the aaa analysis
Rizin report warning messages increasingly when repeated run aaa.
This is due to the fact that analysis information aren't reset between
2 aaa analysis so the blocks analysis isn't deterministic.

The simple 'hack' is to force the fcns analysis data to be reset.

I add a simple test to verify that the second analysis isn't overriding
the first one.