Use RZ_LOG_INFO instead of RZ_LOG_WARN for hash/gnu_hash sections
Right now we don't correctly handle some ELF cases where the gnu_hash
section is ill-formed to allow the LD loader to skip it quickly (e.g.
when no symbols are defined in the binary). In such cases, we should not
bombard the user with warnings that may confuse them, considering the
situation has just improved compared to the last release.

However, we want in the future to properly count dynamic symbol entries
by using relocations. Until then though, the users don't want to see
warnings even for regular binaries.