Use a stack to store the trees/input while applying command edits (#1806)
* Use a stack to store the trees/input while applying command edits

Command edits are performed when there is a need to rewrite part of the
user input, in order to support things like command substitution and
also rewriting one command in terms of another (e.g. @@c: is rewritten
in terms of @@=).

It can happen that multiple nested edits are necessary in execute a
script, thus it is necessary to maintain a stack of saved things,
instead of just one element.