Convert aae to newshell and remove /re
The functionality to analyze only for references that point to a
specific addr, available as `aae <size> <addr>` and `/re` was untested
and at least half broken, so it was removed.
The rest of the code has also been refactored to use less strings and
global state (no core->offset).