Remove `la` instruction.
`la` is a mnemonic for `addi`. I could not find a single occurrence in any of the test binaries of it.
The assembler always resolves the `la` mnemonic into `addi`.
I couldn't find a documented bit difference between `la` and `addi`. Because the instruction can not be tested it is removed.