Remove SPR instructions which are not supported by QEMU or not traced yet.
For most set/read SPR instructions QEMU segfaults.
Those which don't cause segfaults are not traced in QEMU currently.
Because they are barely used anyways I removed them (instead if imeplenting the tracing in QEMU which would have cost too much time for me currently).

In case of SPR 1 (xer), 8 (lr) and 9 (ctr) the assembler resolves them to their mnemonics (mtxer, mtlr etc.). This means the code here is never reached.
To test the get_xer code MFXER was added again. The rz-tracetests will fail for this instructions (due to missing ca32, ov32). But this case is covert in an issue.