Expose rz_core_prevop_addr* as RZ_API + docs (#3036)
* Expose rz_core_prevop_addr* as RZ_API + docs

This was sort of a regression from
59b38e6efaf00b9b9869e0ec5baba4f1b9605f37, where it was switched to
RZ_IPI in the header to be consistent with the source. However we need
RZ_API because it is used at least in Cutter.

* Move rz_core_prevop_addr* to canalysis.c

Since it is api, it is not bound to TUI only, so fits better in generic
core analysis code.