ci: build portable Windows
Before this patch only the vs2019_static Windows build was compiled in
portable mode, however the Windows installer that is part of the release
is based on the clang_cl build. That meant a user installing Rizin on
Windows would get wrong paths when doing `rizin -H`, as those paths
would reference the CI user paths.