Remove syncing dbg->bits -> asm.bits
As already mentioned in 8681cdc43b708eeda28bf3920455fc714cb9c71c,
the deprecated dbg->bits is set and used quite randomly, for example on
macOS/arm64, disassembling sometimes sets it to 16, then stepping with
dbg.follow enabled syncs that value back to asm.bits, which makes no
sense and breaks further disassembly.
So in light of a future removal of dbg->bits, we remove this usage of
it to fix the issue.