Catch rz_asm_tokenize_asm_string() returning NULL
rz_asm_tokenize_asm_string() can return NULL for example if the given
string is empty, so this has to be caught.
In rz_asm_colorize_asm_str(), we could fallback to returning an empty
string, but since this function allocates on the heap, it may return
NULL in any case, so we have to adjust all of its callers anyway and
that means we can just pass the NULL from rz_asm_tokenize_asm_string().

`rz -Qc 'pd 2 @e:asm.pseudo=1' test/bins/elf/mipsloop` shows the initial
issue.