RzEgg: return error codes on failures (#3484)
* Return error codes on failures

Many rz_egg_* APIs can fail, but failures were previously ignored by rz-gg (or just logged).
Exit on failure, and attempt to detect more possible failures.

* Fix rz_egg_padding error format

* Mark failing tests as broken

This is not a perfect solution. the problem is that in big-endian environments, immediates
are encoded according to the host endianness for some reason.
In any case, add a test that the compilation does not fail completely.