Remove warning about missing calling convention
This warning turned out to be more misleading and spammy than actually
useful. For example, it speaks of a "file" when loading just a malloc,
and many architectures simply do not have any standardized calling
conventions, like 6502 or gb, in which case the analysis is fine without
any loaded cc info.