Remove check for theoretically unreachable effects.
The assumptions made in this line doesn't need to be true.
For Hexagon every write to a register first happens to a temporary
register file. Once the whole packet is checked and is valid,
the operations are performed.
The last step is to sync the the temporary register file with the default one.