librz: rewrite base64 code
The new Base64 code is fixing the previous bugs which were illustrated
with unit tests. Documentation is added and as a "breaking" API
change, the input and output parameters are allowed to be NULL
consistently (not just some, but all.) The rest of the behavior is the
same, save for the bugs that were fixed, and the decoding function
which is now even laxer in what it is able to decode (which seemed to
be the original intention but had a bug as demonstrated by a unit
test.)

An integer overflow bug, in the calculation of the string length, is
also fixed.