Fix #3965 (#3969)
This fixes a reachable assert(). The operands used
to calculate the jump target, used whatever was in the
union of the operand value field.

This also means that the previous calculation of
jump targets were simply incorrect.
It nontheless lead to the right results because
MEMDISP() seemed to be 0 all the time.