Make rz_flag_set_next() try the addr before incremental numbers as suffix (#3986)
In binaries with a large number of flags of the same name, like in macho
with many relocs to the same symbol, using an incremental number as
suffix to generate unique flag names leads to quadratic runtime. We can
avoid this by first trying to append the address, which as a nice side
effect also results in more meaningful names.