Don't match numbers with a '.', '*' and '\w' prefix.
The exclusion pattern [^\w.*] excludes all numbers with a '.' and '*' prefix.
This is only done because the previous parsing did it like this.
And not doing it, breaks substitutions from before. E.g. when addresses,
were replaced with flag names which include numbers (case labels like: case.<switch-address>.<case-number>).
See: https://github.com/rizinorg/rizin/issues/4238 for more details of this problem.