Adjust OpenBSD code to set the process state to be less wrong.
Use the p_stat of the main thread to decide in which state the process is.
For multi-threaded applications this is still not quite correct but at
least this compiles on OpenBSD-current.

See also https://github.com/openbsd/ports/commit/5d4f90ee105757eda1cd9b31108c3ef6df67c48b

Fix for #4576