Rewrite string detection.
Short summary: It is rather straight forward. The detection function gets a buffer it iterates twice over the buffer ('O(2n)') and
decodes at each index a code point for each possible encoding and checks it's printability ('O(|Encodings| * log(I)): I is number of invalid code point ranges.').
If the decoding gives a _printable_ code point, it gets a score of the number of bytes it uses (4 if UTF-32, 2-4 for UTF-16, 1 for IBM037 etc.).
If there is no valid decoding of the bytes, the score is '0.0'.
This is done for every offset.

Now we iterate backwards over the buffer and accumulate the scores, except they have a score of '0.0'.
This means that strings of printable characters have a high score at the beginning of the string (namely the number of consecutive printable characters) and a low score at the end.