Remove unused min_uni_blocks setting and move the number into a constant.
The setting was somewhat related to false positive string detection.
It seemed to be the value how many blocks the characters of a string
can be in. But it is too ineffecient to scan this in this implementation.
Also not a setting which should be exposed to the user in this way.