Fix naming of 'RzBinMap->paddr' to 'offset' to be consistent with common binary formats.
This name change is an attempt to be more precise what the member actually does.

The paddr member was effectively used as an offset into the binary file.
It holds the offset where the map starts.

The name "paddr" is commonly used for "physical address" though.
But a file offset is not the same as a physical address in most contexts.

A physical address commonly refers to a memory space which has some kind of address
translation unit, which manages the memory space by giving out virtual addresses to
memory users.
The physical address refers then to the actual offset into the memory space
(while the virtual addresses don't need fixed to the same physical memory location).

In the context of RzBinMap the paddr is an offset into the binary file.
Not into a memory space. One could argue that the binary file itself is a "physical memory space",
but this is a rather uncommon interpretation.

More so, if RzBinMap is used in context of binary formats which have the concept of mapping (ELF, PE etc)
this naming can be a source of confusion. Because physical address in those binary formats refers to
actual physical addresses as described above. And offsets into files are referred as offset.
For example ELF has for program headers: p_paddr, p_vaddr and p_offset.
For sections only: sh_offset, sh_addr (sh_addr is virtual or physical depending what the machine supports).

The name change from paddr to offset attempts to fix this potential source of confusion.